sorta working v2 support
throws java.lang.ArrayIndexOutOfBoundsException for no decipherable reason
